# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "PRIN144-Final-Exam",
      "version": "1.0",
      "description": "Description",
      "contact": {
        "email": "syuto7702@gmail.com",
        "name": "Syuto Ramos"
      }
    },
    "externalDocs": {
      "url": "swagger.json",
      "description": "swagger.json"
    },
    "servers": [
      {
        "url": "https://prin-144-final-exam-eight.vercel.app/",
        "variables": {},
        "description": "Production"
      }
    ],
    "paths": {
      "/student": {
        "get": {
          "summary": "Returns all tasks.",
          "description": "Returns all tasks.",
          "operationId": "",
          "responses": {
            "200": {
              "description": "Task Found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/studentGetResponse"
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Task not found"
            },
            "default": {
              "description": "Default error sample response"
            }
          },
          "tags": [
            "Tasks"
          ]
        },
        "post": {
          "summary": "Creates a task",
          "description": "Creates a task",
          "operationId": "",
          "requestBody": {
            "description": "Post description",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/studentCreateRequest"
                }
              }
            }
          },
          "responses": {
            "default": {
              "description": "Default error sample response"
            }
          },
          "tags": [
            "Tasks"
          ]
        }
      },
      "/student{id}": {
        "get": {
          "summary": "Returns a task by specified id.",
          "description": "Returns a task by specified id.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The id of the task.",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "operationId": "",
          "responses": {
            "200": {
              "description": "Task Found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/studentGetResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Task not found"
            }
          },
          "tags": [
            "Tasks"
          ]
        },
        "put": {
          "summary": "Updates a task by specified id.",
          "description": "Updates a task by specified id.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The id of the task.",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "operationId": "",
          "requestBody": {
            "description": "Update description",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/studentUpdateRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Task Found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/studentGetResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Task not found"
            },
            "default": {
              "description": "Default error sample response"
            }
          },
          "tags": [
            "Tasks"
          ]
        },
        "delete": {
          "summary": "Deletes a task by specified id.",
          "description": "Deletes a task by specified id.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The id of the task.",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "operationId": "",
          "responses": {
            "200": {
              "description": "Task successfully deleted"
            },
            "404": {
              "description": "Task not found"
            }
          },
          "tags": [
            "Tasks"
          ]
        }
      }
    },
    "components": {
      "schemas": {
        "studentCreateRequest": {
          "type": "object",
          "required": [
            "GivenName",
            "FamilyName",
            "YearLevel",
            "Course",
            "Section"
          ],
          "properties": {
            "GivenName": {
              "type": "string"
            },
            "FamilyName": {
              "type": "string"
            },
            "YearLevel": {
              "type": "integer"
            },
            "Course": {
              "type": "string",
              "enum": [
                "BSIT",
                "BSCS",
                "BSCoEn"
              ]
            },
            "Section": {
              "type": "integer"
            }
          }
        },
        "studentUpdateRequest": {
          "type": "object",
          "properties": {
            "GivenName": {
              "type": "string"
            },
            "FamilyName": {
              "type": "string"
            },
            "YearLevel": {
              "type": "integer"
            },
            "Course": {
              "type": "string",
              "enum": [
                "BSIT",
                "BSCS",
                "BSCoEn"
              ]
            },
            "Section": {
              "type": "integer"
            }
          }
        },
        "studentGetResponse": {
          "type": "object",
          "properties": {
            "GivenName": {
              "type": "string"
            },
            "FamilyName": {
              "type": "string"
            },
            "YearLevel": {
              "type": "integer"
            },
            "Course": {
              "type": "string",
              "enum": [
                "BSIT",
                "BSCS",
                "BSCoEn"
              ]
            },
            "Section": {
              "type": "integer"
            }
          }
        }
      }
    }
  }